# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shelley Point Golf Club Scorecard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Hole {
  number: number;
  par: number;
  /** Distance in metres from yellow (men) tee */
  distanceMen: number;
  /** Distance in metres from blue (senior/ladies) tee */
  distanceLadies: number;
  strokeIndex: number;
  name: string;
  description: string;
  teeLat: number;
  teeLng: number;
  pinLat: number;
  pinLng: number;
}

export interface Player {
  name: string;
  handicap: number;
}

export interface ShotRecord {
  club: string;
  /** GPS distance to pin at time of shot in metres */
  distanceToPin: number;
  ballLat?: number | null;
  ballLng?: number | null;
}

export interface HoleScore {
  holeNumber: number;
  scores: (number | null)[];
  /** Shot records per player (same order as players array) */
  shots?: ShotRecord[][];
}

/**
 * stroke = Stroke Play, stableford = Stableford Points, match = Match Play
 */
export type GameFormat = (typeof GameFormat)[keyof typeof GameFormat];

export const GameFormat = {
  stroke: "stroke",
  stableford: "stableford",
  match: "match",
} as const;

export interface Scorecard {
  id: number;
  date: string;
  gameFormat: GameFormat;
  /** Number of holes in this round */
  holesCount: 9 | 18;
  /** Starting hole number (1 = Front 9, 10 = Back 9) */
  startingHole: 1 | 10;
  players: Player[];
  holeScores: HoleScore[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateScorecardRequest {
  date: string;
  gameFormat: GameFormat;
  /** Number of holes to play (9 or 18) */
  holesCount?: 9 | 18;
  /** Starting hole number (1 = Front 9, 10 = Back 9) */
  startingHole?: 1 | 10;
  /**
   * @minItems 1
   * @maxItems 4
   */
  players: Player[];
}

export interface UpdateScorecardRequest {
  holeScores: HoleScore[];
  players?: Player[];
  gameFormat?: GameFormat;
}