# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shelley Point Golf Club Scorecard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scorecard
    description: Golf scorecard operations
  - name: course
    description: Course information
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /course/holes:
    get:
      operationId: getCourseHoles
      tags: [course]
      summary: Get all course holes
      responses:
        "200":
          description: Course holes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hole"

  /scorecards:
    get:
      operationId: listScorecards
      tags: [scorecard]
      summary: List all scorecards
      responses:
        "200":
          description: List of scorecards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scorecard"
    post:
      operationId: createScorecard
      tags: [scorecard]
      summary: Create a new scorecard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScorecardRequest"
      responses:
        "201":
          description: Created scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scorecard"

  /scorecards/{id}:
    get:
      operationId: getScorecard
      tags: [scorecard]
      summary: Get a scorecard by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scorecard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateScorecard
      tags: [scorecard]
      summary: Update a scorecard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScorecardRequest"
      responses:
        "200":
          description: Updated scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scorecard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteScorecard
      tags: [scorecard]
      summary: Delete a scorecard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Hole:
      type: object
      properties:
        number:
          type: integer
        par:
          type: integer
        distanceMen:
          type: integer
          description: Distance in metres from yellow (men) tee
        distanceLadies:
          type: integer
          description: Distance in metres from blue (senior/ladies) tee
        strokeIndex:
          type: integer
        name:
          type: string
        description:
          type: string
        teeLat:
          type: number
        teeLng:
          type: number
        pinLat:
          type: number
        pinLng:
          type: number
      required:
        - number
        - par
        - distanceMen
        - distanceLadies
        - strokeIndex
        - name
        - description
        - teeLat
        - teeLng
        - pinLat
        - pinLng

    Player:
      type: object
      properties:
        name:
          type: string
        handicap:
          type: integer
      required:
        - name
        - handicap

    ShotRecord:
      type: object
      properties:
        club:
          type: string
        distanceToPin:
          type: integer
          description: GPS distance to pin at time of shot in metres
        ballLat:
          type: number
          nullable: true
        ballLng:
          type: number
          nullable: true
      required:
        - club
        - distanceToPin

    HoleScore:
      type: object
      properties:
        holeNumber:
          type: integer
        scores:
          type: array
          items:
            type: integer
            nullable: true
        shots:
          type: array
          description: Shot records per player (same order as players array)
          items:
            type: array
            items:
              $ref: "#/components/schemas/ShotRecord"
      required:
        - holeNumber
        - scores

    GameFormat:
      type: string
      enum: [stroke, stableford, match]
      description: "stroke = Stroke Play, stableford = Stableford Points, match = Match Play"

    Scorecard:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        gameFormat:
          $ref: "#/components/schemas/GameFormat"
        holesCount:
          type: integer
          enum: [9, 18]
          description: Number of holes in this round
        startingHole:
          type: integer
          enum: [1, 10]
          description: Starting hole number (1 = Front 9, 10 = Back 9)
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        holeScores:
          type: array
          items:
            $ref: "#/components/schemas/HoleScore"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - date
        - gameFormat
        - holesCount
        - startingHole
        - players
        - holeScores
        - createdAt
        - updatedAt

    CreateScorecardRequest:
      type: object
      properties:
        date:
          type: string
        gameFormat:
          $ref: "#/components/schemas/GameFormat"
        holesCount:
          type: integer
          enum: [9, 18]
          default: 18
          description: Number of holes to play (9 or 18)
        startingHole:
          type: integer
          enum: [1, 10]
          default: 1
          description: Starting hole number (1 = Front 9, 10 = Back 9)
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
          minItems: 1
          maxItems: 4
      required:
        - date
        - gameFormat
        - players

    UpdateScorecardRequest:
      type: object
      properties:
        holeScores:
          type: array
          items:
            $ref: "#/components/schemas/HoleScore"
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        gameFormat:
          $ref: "#/components/schemas/GameFormat"
      required:
        - holeScores